proximity doesn't need to do er io
